coll: merge barrier dissemination to the high radix dissemination
Make the optimized version of dissemination algorithm to be a
special case of the high radix dissemination algorithm.
