--- src/mpi/coll/barrier/barrier_intra_dissemination.c
@@ -1,50 +1,0 @@
-/*
- * Copyright (C) by Argonne National Laboratory
- *     See COPYRIGHT in top-level directory
- */
-
-#include "mpiimpl.h"
-
-/* Algorithm: MPI_Barrier
- *
- * We use the dissemination algorithm described in:
- * Debra Hensgen, Raphael Finkel, and Udi Manbet, "Two Algorithms for
- * Barrier Synchronization," International Journal of Parallel
- * Programming, 17(1):1-17, 1988.
- *
- * It uses ceiling(lgp) steps. In step k, 0 <= k <= (ceiling(lgp)-1),
- * process i sends to process (i + 2^k) % p and receives from process
- * (i - 2^k + p) % p.
- */
-int MPIR_Barrier_intra_dissemination(MPIR_Comm * comm_ptr, MPIR_Errflag_t * errflag)
-{
-    int size, rank, src, dst, mask, mpi_errno = MPI_SUCCESS;
-    int mpi_errno_ret = MPI_SUCCESS;
-
-    size = comm_ptr->local_size;
-    rank = comm_ptr->rank;
-
-    mask = 0x1;
-    while (mask < size) {
-        dst = (rank + mask) % size;
-        src = (rank - mask + size) % size;
-        mpi_errno = MPIC_Sendrecv(NULL, 0, MPI_BYTE, dst,
-                                  MPIR_BARRIER_TAG, NULL, 0, MPI_BYTE,
-                                  src, MPIR_BARRIER_TAG, comm_ptr, MPI_STATUS_IGNORE, errflag);
-        if (mpi_errno) {
-            /* for communication errors, just record the error but continue */
-            *errflag =
-                MPIX_ERR_PROC_FAILED ==
-                MPIR_ERR_GET_CLASS(mpi_errno) ? MPIR_ERR_PROC_FAILED : MPIR_ERR_OTHER;
-            MPIR_ERR_SET(mpi_errno, *errflag, "**fail");
-            MPIR_ERR_ADD(mpi_errno_ret, mpi_errno);
-        }
-        mask <<= 1;
-    }
-
-    if (mpi_errno_ret)
-        mpi_errno = mpi_errno_ret;
-    else if (*errflag != MPIR_ERR_NONE)
-        MPIR_ERR_SET(mpi_errno, *errflag, "**coll_fail");
-    return mpi_errno;
-}
