+++ src/mpi/coll/barrier/barrier_intra_k_dissemination.c
@@ -5,6 +5,42 @@
 
 #include "mpiimpl.h"
 
+/* Algorithm: MPI_Barrier
+ *
+ * We use the dissemination algorithm described in:
+ * Debra Hensgen, Raphael Finkel, and Udi Manbet, "Two Algorithms for
+ * Barrier Synchronization," International Journal of Parallel
+ * Programming, 17(1):1-17, 1988.
+ *
+ * It uses ceiling(lgp) steps. In step k, 0 <= k <= (ceiling(lgp)-1),
+ * process i sends to process (i + 2^k) % p and receives from process
+ * (i - 2^k + p) % p.
+ */
+int MPIR_Barrier_intra_dissemination(MPIR_Comm * comm_ptr, MPIR_Errflag_t * errflag)
+{
+    int size, rank, src, dst, mask, mpi_errno = MPI_SUCCESS;
+    int mpi_errno_ret = MPI_SUCCESS;
+
+    size = comm_ptr->local_size;
+    rank = comm_ptr->rank;
+
+    mask = 0x1;
+    while (mask < size) {
+        dst = (rank + mask) % size;
+        src = (rank - mask + size) % size;
+        mpi_errno = MPIC_Sendrecv(NULL, 0, MPI_BYTE, dst,
+                                  MPIR_BARRIER_TAG, NULL, 0, MPI_BYTE,
+                                  src, MPIR_BARRIER_TAG, comm_ptr, MPI_STATUS_IGNORE, errflag);
+        MPIR_ERR_COLL_CHECKANDCONT(mpi_errno, *errflag);
+        mask <<= 1;
+    }
+
+    if (mpi_errno_ret)
+        mpi_errno = mpi_errno_ret;
+    else if (*errflag != MPIR_ERR_NONE)
+        MPIR_ERR_SET(mpi_errno, *errflag, "**coll_fail");
+    return mpi_errno;
+}
 
 /* Algorithm: high radix dissemination
  * Similar to dissemination algorithm, but generalized with high radix k
@@ -27,6 +63,10 @@ int MPIR_Barrier_intra_k_dissemination(MPIR_Comm * comm, int k, MPIR_Errflag_t *
 
     if (nranks < k)
         k = nranks;
+
+    if (k == 2) {
+        return MPIR_Barrier_intra_dissemination(comm, errflag);
+    }
 
     /* If k value is greater than the maximum radix defined by MAX_RADIX macro,
      * we allocate memory for requests here. Otherwise we use the requests defined
