+++ test/mpi/maint/coll_cvars.txt
@@ -308,7 +308,6 @@ algorithms:
         inter-nonblocking:
             sched_bcast
         intra-blocking:
-            dissemination
             k_dissemination
                 .MPIR_CVAR_BARRIER_DISSEM_KVAL=2,3,4,8
             recexch
